ARC GIS User Interface
In this updated version, I have used the ipywidgets library to create a user interface with input fields for the table name, field name, and field value. The user can enter these values and click the "Visualize and Manipulate Data" button to visualize the data on a map and manipulate it by adding a new field. The map will be displayed in the notebook.

Please note that you need to install the ipywidgets library if you haven't already. You can install it using pip: